Chadzirobana HW #1

+++ src/main/java/ge/itstep/demo/dto/TaskDTO.java
@@ -1,4 +1,6 @@
 package ge.itstep.demo.dto;
 
-public record TaskDTO(int id, String task, String note, String dueDate, boolean completed) {
+import java.util.Date;
+
+public record TaskDTO(Long id, String task, String note, String dueDate, boolean completed) {
 }
